Marriage Abroad improvement: Finland (2 outcomes)
In order to simplify the maintenance of the marriage-abroad code,
we are going to redefine the structure of some templates.

Here we are going to define only 2 templates for Finland:
- 1 for uk-resident users
- 1 for non-uk resident users (i.e. resident in Finland on in
another country)

This template structure is different from the one we already defined
from 2 outcomes countries. There the template definition is
determined by the gender of the partner, here it is defined by the
legal residency of the user.

We are therefore applying some changes to make this new behaviour
available. I.e. create:
- a title template partial for Finland
- a uk-resident template partial
- a non-uk-resident template partial
- an outcome template `outcome_marriage_abroad_in_finland.rb`, which
selects the right partial on the basis of the
`calculator.resident_of_uk?` value.

We also modified the marriage-abroad flow
(`lib/smart_answer_flows/marriage-abroad.rb`) in order to get the
outcome `:outcome_marriage_abroad_in_finland` after the
`:legal_residency?` node.

The tests related to Finland were updated appropriately.