Move Tax Year logic into a more generic factory method
The `TaxYear` class contains logic for creating a `YearRange` with
fixed start day and month. This is a good abstraction, but could be
abstracted further so we can apply this logic to other year ranges,
such as the agricultural holiday period (which runs from 1 Oct).

This change extracts this logic into a class factory method. The tax
year unit tests are left in place as good coverage for this factory
method.