Use the unpublish type of vanish for some retiring
When retiring a smart answer, and replacing it with a simple smart
answer, use an unpublishing type of "vanish" for unpublishing the
smart answer.

Previously, unpublishing the smart answer with a type of "gone" would
leave the route in place for the smart answer, and being a more
specific route than the simple smart answer, this would take
precedence.

Until recently, unpublishing with a type of vanish wouldn't have
removed the route from the router, but this should now happen.